fix: 修复机器人批量压测失败问题，支持大规模并发登录  - robots 批量登录数量计算：count/threadCnt 整数除法导致非整倍数时少创建机器人，改为 double 除法后 ceil - 网络层：连接关闭时清理发送队列中指向该 socket 的包，避免 fd 被新连接复用后旧包串包被丢弃 - 登录服：新增第三方验证 10 秒超时自愈机制，验证请求丢失时踢掉卡住的账号、删除登录锁并通知客户端，机器人断线后自动重试 - 登录服：超时踢人时显式删除 redis 登录锁，防止残留锁导致重试被误判"账号在线" - web/member_login_t.php：本地测试模式改为任意账号+正确密码即通过，支持批量压测账号 - run-all.sh / run-allinone.sh：支持 release 参数启动 Release 版；PHP 内置服务器启用 PHP_CLI_SERVER_WORKERS=4 多进程并行，避免批量验证请求排队 - stop-all.sh：同时停止 Debug 和 Release 两个版本的服务进程

+++ src/tools/robots/robot_console_login.cpp
@@ -56,8 +56,8 @@ void RobotConsoleLogin::HandleLoginEx(std::vector<std::string>& params) const
     auto threadCnt = pAppConfig->LogicThreadNum;
     threadCnt = threadCnt <= 0 ? 1 : threadCnt;
 
-    // 每个线程中robot数量
-    int perThreadRobotCnt = static_cast<int>(std::ceil(count / threadCnt));
+    // 每个线程中robot数量（注意整数除法，需先转 double 再 ceil，否则非整倍数时数量会少）
+    int perThreadRobotCnt = static_cast<int>(std::ceil(static_cast<double>(count) / threadCnt));
     perThreadRobotCnt = perThreadRobotCnt < 1 ? 1 : perThreadRobotCnt;
 
     for (int i = 0; i < threadCnt; i++)
